requirements.txt
numpy>=1.20.0
pettingzoo>=1.24.0
gymnasium>=0.26.0
networkx>=2.6.0
pyyaml>=5.4.0